Analog engine contrast, strip colour bug, +60px window, modal preset sheet
AnalogEngine. Two findings, measured through the FULL path - channel, dry
bus, master - which nothing had been measured through before: runConsole and
the level-matching test both stopped at the bus.

  1. With the master included the engine LOST level: 0.8 dB on CLEAN up to
     3.5 dB on TRANSFORMER. Switching it on made the mix quieter, so every
     A/B was partly a loudness comparison. Trims are re-solved against the
     full path and land at 0.00 dB.
  2. curveBlend, not drive, is what sets how much character it has. Raising
     pairDrive does almost nothing - the channel and bus share it and the
     pair cancels; THD did not even rise monotonically with it. The blend
     changes the shape of the transfer, which the inverse cannot undo.

Blends raised to roughly double the harmonic content at full amount:
CLEAN 0.22 -> 0.54%, BRITISH 2.80 -> 6.64%, AMERICAN 6.93 -> 9.97%,
TRANSFORMER 1.65 -> 4.68%, MODERN 2.06 -> 3.90%. AMERICAN sits at 0.66
rather than 0.744 because past that the inverse clamp engages and THD folds
back instead of rising with level.

outputTrim now applies on the inverse side and at the master only. On the
forward channel stage it sat between the channel's transfer and the bus's
inverse, so the bus inverted a signal that had been scaled since and the
pair stopped cancelling - 0.045% THD on a single channel that is
transparent by construction. Back to 0.001%.

FX strip colours. The nodes read their colour in sectionAccent at the moment
they are BUILT, and they are built in the panel's constructor, from addCard
and from setChainOrder - all before any config exists, where sectionAccent
returns the panel's own accent. setUIConfig stored the config without
rebuilding them, so all eight nodes stayed the same blue until something
unrelated rebuilt them; clicking one reordered the chain, which called
setChainOrder, which refreshed them, and the strip snapped to its real
colours at once. Noticed while writing the test: in this build
resolveUiConfigFile only probes the bundle - the cwd and source-tree
candidates are behind JUCE_DEBUG || PX3_DEBUG_PANEL - so the editor under
test loads no config at all, and the test drives FxPanel directly.

Window is 60px taller for the taller FX cards, and all 60 go to the panels.
The keyboard is sized as a fraction of the window, so its fraction is
re-based 0.15 -> 0.14 to keep it at the same 106px; the header is fixed
already.

Preset sheet is now modal in behaviour, not just in appearance.
paintOverChildren dimmed the editor while every knob, card and key
underneath stayed live, so a click that missed the sheet edited the patch
you were browsing away from. A scrim across the editor sits under the sheet
and forwards its mouse events to the editor, which is where the sheet's own
title-bar drag and click-outside-to-close already live.

+++ Source/UI/FxSignalFlow.cpp
@@ -157,7 +157,7 @@ void FxSignalFlow::mouseDrag(const juce::MouseEvent& event)
 
     auto& dragged = current[static_cast<std::size_t>(draggedIndex)];
     const auto newX = juce::jlimit(getLocalBounds().toFloat().getX(),
-                                   juce::jmax(0.0f, getWidth() - dragged.getWidth()),
+                                   juce::jmax(0.0f, static_cast<float>(getWidth()) - dragged.getWidth()),
                                    static_cast<float>(event.getPosition().x - dragOffsetX));
     dragged.setX(newX);
